Moved to keras subfolder, fix some stuff
